created Login function
