+++ app/src/main/java/com/group80/uoftinder/Login.java
@@ -1,0 +1,70 @@
+package com.group80.uoftinder;
+
+import static android.content.ContentValues.TAG;
+
+import android.os.Bundle;
+import android.util.Log;
+import android.widget.EditText;
+import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+
+public class Login extends AppCompatActivity {
+    private FirebaseAuth mAuth;
+
+    String email = ((EditText) findViewById(R.id.email)).getText().toString().trim();
+    String password = ((EditText) findViewById(R.id.password)).toString().trim();
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.loginview);
+
+        mAuth = FirebaseAuth.getInstance(); // initialize the Firebase Auth
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        // Check if user is signed in (non-null) and update UI accordingly.
+        FirebaseUser currentUser = mAuth.getCurrentUser(); // this is not null ONLY if you have logged in
+        /* To log-in, use the following
+         * mAuth.signInWithEmailAndPassword("demo@email.com", "password");
+         * To get user uid, after you have signed in, do the following
+         * mAuth.getCurrentUser().getUid();
+         */
+        if(currentUser != null){
+            currentUser.reload();
+        }
+        else {
+            mAuth.signInWithEmailAndPassword(email, password)
+                    .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
+                        @Override
+                        public void onComplete(@NonNull Task<AuthResult> task) {
+                            if (task.isSuccessful()) {
+                                // Sign in success, update UI with the signed-in user's information
+                                Log.d(TAG, "signInWithEmail:success");
+                                FirebaseUser user = mAuth.getCurrentUser();
+                                mAuth.getCurrentUser().getUid();
+
+                            } else {
+                                // If sign in fails, display a message to the user.
+                                Log.w(TAG, "signInWithEmail:failure", task.getException());
+                                Toast.makeText(Login.this, "Authentication failed.",
+                                        Toast.LENGTH_SHORT).show();
+                            }
+                        }
+                    });
+            }
+
+
+        }
+
+    }
